backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Authentication & Security (Enhanced)
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
cryptography>=41.0.0
email-validator==2.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# AI and ML Libraries (Enhanced)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
llama-index>=0.9.30
llama-index-agent-openai>=0.1.0
llama-index-callbacks-langfuse>=0.1.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
python-magic==0.4.27
textract==1.6.5

# Email
fastapi-mail==1.4.1

# Security & Compliance
python-jose[cryptography]==3.3.0
Pillow>=10.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
celery==5.3.4
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.7.1

# Production dependencies
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Data validation
email-validator==2.1.0
